Refactor: move enums into shared folder.
