+++ src/model/enums/mod.rs
@@ -1,0 +1,14 @@
+//! # EVE ESI Enums
+//!
+//! Defines shared enums used across ESI models
+//!
+//! # Modules
+//! - [`asset`]: Asset-related enums
+//! - [`character`]: Character-related enums
+//! - [`corporation]: Corporation-related enums
+//! - [`notification`]: Notification-related enums
+
+pub mod asset;
+pub mod character;
+pub mod corporation;
+pub mod notification;
